src/requirements.txt
pandas==2.2.2
openai==1.50.0
websocket-client==1.8.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
rich==13.7.1

structlog>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyarrow>=14.0.0
PyYAML>=6.0.0
